requirements.txt
Quart-Discord==2.1.4
quart==0.16.1
python-dotenv==0.19.2
